+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+Changed
+
+- `_split_on_dot` is no longer in base parser (now in a utils file). Also, the default 
+  behavior is now to split on every dot and split on only the last dot if specified
+  
+Fixed
+
+- `ConfigFile` was trying to use `_split_on_dot`, but it no longer inherited from base parser.
+
 ## 0.3.1 - 2020-01-17
 
 Fixed