+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.3
+
+Fixed
+
+- `TypeError` when setting the key/value pairs since configparser requires the option 
+   to be a string. The parser now just converts the value to a string if it is not one
+   and then adds the key/value pair. It would still be parsed correctly when retrieving
+   it.
+
 ## 0.3.2 - 2020-01-17
 
 Changed